backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
black==23.11.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.6
emoji==2.14.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.104.1
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
flake8==6.1.0
Flask==3.1.0
fonttools==4.58.5
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.2.0
gradio==5.16.1
gradio_client==1.7.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.20.3
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.3.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.0.350
langchain-community==0.0.5
langchain-core==0.1.23
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.24.4
openai==1.3.0
opencv-python==4.8.1.78
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdf2image==1.16.3
pdfminer.six==20250506
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.3.6
plotly==6.2.0
plotly-express==0.4.1
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.3.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.10.1
pydantic_core==2.14.1
pydub==0.25.1
pyflakes==3.1.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rubicon-objc==0.5.0
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.5
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.2
typer==0.15.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritools==5.0.0
urlextract==1.9.0
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wikipedia==1.4.0
yarl==1.20.1
